# docs/swagger.yaml
basePath: /api
definitions:
...
  req.RequestAddStatusBooking:
    properties:
      description:
        type: string
      status_code:
        type: string
      status_name:
        type: string
    required:
    - description
    - status_code
    - status_name
    type: object
...
  req.RequestUpdateAccount:
    properties:
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      password:
        type: string
      status:
        type: integer
    type: object
...
  res.Response:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: https://siphoria-api-production.up.railway.app
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HOTEL BOOKING API
  version: "1.0"
paths:
  /admin/accept/:id:
    patch:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: Accept
      tags:
      - admin-service
  /admin/accounts:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: 'Get account admin (Sort: pass sort=field&order=desc, Filter: field_you_want_to_pass=value,
        Paging: page=0&offset=3, Search: search=8)'
      tags:
      - admin-service
  /admin/create-account:
    post:
      consumes:
      - application/json
      parameters:
      - description: status booking
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestAddStatusBooking'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: create account by admin
      tags:
      - admin-service
  /admin/hotels:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: Get hotel list
      tags:
      - admin-service
  /admin/payouts/:id:
    patch:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: Approve request payout
      tags:
      - admin-service
...
  /admin/update-account:
    patch:
      consumes:
      - application/json
      parameters:
      - description: staffaccount
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestUpdateAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: update account
      tags:
      - admin-service
  /admin/update-cmsrate/:id:
    patch:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: Update commission rate hotel
      tags:
      - admin-service
  /admin/update-rating/:id:
    patch:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: Update rating hotel
      tags:
      - admin-service
...
swagger: "2.0"